Initial implementation of GUI plus CLI
